create read delete réunion

+++ src/main/java/com/cesi/projetJavaSpringBoot/repositories/ReunionRepository.java
@@ -3,5 +3,5 @@ package com.cesi.projetJavaSpringBoot.repositories;
 import com.cesi.projetJavaSpringBoot.entities.Reunions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReunionRepository extends JpaRepository<Reunions, Integer> {
+public interface ReunionRepository extends JpaRepository<Reunions, Long> {
 }
